Fixed and simplified the setting of PROMPT_COMMAND
The #2 instruction told user to set PROMPT_COMMAND with a string that used "proper" double quotes and newer bash subshell syntax, which required three escape sequences, one of them even a triple-escape. And - crucially - at some point quit working. The new instruction uses a string that falls back to old Bash syntax with single quotes and back-ticks, which requires no escaping, is easier to read, and - kind of importantly - works.